requirements.txt
requests
suds-jurko
reportlab
